image_crop/requirements.txt
opencv-python~=4.12.0.88
pillow~=11.3.0
tqdm~=4.67.1